+++ x/auth/keeper/params.go
@@ -1,0 +1,17 @@
+package keeper
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth/types"
+)
+
+// SetParams sets the auth module's parameters.
+func (ak AccountKeeper) SetParams(ctx sdk.Context, params types.Params) {
+	ak.paramSubspace.SetParamSet(ctx, &params)
+}
+
+// GetParams gets the auth module's parameters.
+func (ak AccountKeeper) GetParams(ctx sdk.Context) (params types.Params) {
+	ak.paramSubspace.GetParamSet(ctx, &params)
+	return
+}
